+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Add a convenience function for localized labels in tests (#1268)
 
 ### Changed
+- Switch to the TYPO3 Code of Conduct (#1311)
 - Require oelib >= 4.1.4 (#1281)
 - Clean up the test (#1267)
 - Move more legacy tests to the new testing framework (#1267, #1282, #1290, #1308)